fix: changed two functions to use double instead of float to avoid truncating coordinate precision

+++ MapillarySDK/MapillarySDK/internal/MAPInternalUtils.h
@@ -26,8 +26,8 @@
 
 + (NSString*)appVersion;
 
-+ (float)calculateFactorFromDates:(NSDate*)date date1:(NSDate*)date1 date2:(NSDate*)date2;
-+ (CLLocationCoordinate2D)interpolateCoords:(CLLocationCoordinate2D)location1 location2:(CLLocationCoordinate2D)location2 factor:(float)factor;
++ (double)calculateFactorFromDates:(NSDate*)date date1:(NSDate*)date1 date2:(NSDate*)date2;
++ (CLLocationCoordinate2D)interpolateCoords:(CLLocationCoordinate2D)location1 location2:(CLLocationCoordinate2D)location2 factor:(double)factor;
 + (NSNumber*)calculateHeadingFromCoordA:(CLLocationCoordinate2D)A B:(CLLocationCoordinate2D)B;
 
 + (NSDateFormatter*)defaultDateFormatter;
